requirements.txt
Django
djangorestframework
python-dateutil
httplib2
geopy